completed stringifyJSON
my approach is to identify objects within objects. A recursion identifies an object within an object, it "steps in" to the innerObject then evaluates if the said innerObject contains an object until the object does not contain an object anymore.

Once the loop reaches the innermost object, it starts string conversion and "steps out."